Pere/REINFORCE_FIRST_TRIALS/requirements.txt
ale_py==0.10.1
gymnasium==1.0.0
ipython==8.12.3
matplotlib==3.7.5
numpy==1.24.4
opencv_python==4.10.0.84
Pillow==7.0.0
Pillow==11.0.0
torch==2.4.1
torchsummary==1.5.1
